aider: Add debug print statement for skipped records with empty last names in the `processAndWriteToXlsx` function.
# Aider chat conversation:

USER: looks like the errors are removed. but we only havbe 120 records.

ASSISTANT: To investigate why only 120 records are being processed, we need to examine the `processAndWriteToXlsx` function in the `main_v3.py` file. The issue could be related to the regular expression used for splitting the records, the way records are being filtered out, or an issue with the PDF text extraction itself.

Since the regular expression was recently adjusted, it's possible that the pattern is still not capturing all the records due to inconsistencies in the PDF text. Another possibility is that some records are being inadvertently filtered out by the condition that checks for an empty last name.

To ensure that we are not missing records due to the last name check, we can add a print statement to output records that are being skipped. This will help us identify if there are records with content that are not being processed due to the last name being empty or not matching the expected pattern.

Here are the diffs for adding a debug print statement:

```diff
--- main_v3.py
+++ main_v3.py
@@ ... @@
 def processAndWriteToXlsx():
     ...
     for record in records:
         if record.strip() == '':
             continue
         recordData = prepareRecordForCsv(record)
-        if recordData['LastName'] != '':
+        if recordData['LastName']:
             ws.append([
                 ...
             ])
+        else:
+            print(f"Skipped record due to empty last name: {record}")
     print(len(records))
     wb.save('inmate_records.xlsx')
```

This change will print out the content of records that have an empty last name after being processed by `prepareRecordForCsv`. This can help us identify if there are records that are not being formatted correctly or if there are additional patterns in the PDF text that we need to account for.

Please run the updated code and check the output to see if any records are being skipped and why. If you find that records are being skipped due to issues with the regular expression or the data extraction, please provide examples of the skipped records so that I can further refine the code.